requirements.txt
streamlit                  == 1.11.0
lxml                       == 4.9.1
beautifulsoup4             == 4.11.1
pyyaml                     == 6.0
bcrypt                     == 3.2.2
PyJWT                      == 2.4.0
extra-streamlit-components == 0.1.55